Event service test is not working
