+++ components/event-service/internal/events/bus/send.go
@@ -8,6 +8,8 @@ import (
 	"net/http"
 	"net/url"
 
+	v2 "github.com/kyma-project/kyma/components/event-service/internal/events/bus/v2"
+
 	"github.com/kyma-project/kyma/components/event-service/internal/events/api"
 	"github.com/kyma-project/kyma/components/event-service/internal/httpconsts"
 	"github.com/kyma-project/kyma/components/event-service/internal/httptools"
@@ -34,7 +36,10 @@ func SendEvent(req interface{}, traceHeaders *map[string]string,
 		return nil, err
 	}
 
-	reqURL, err := url.ParseRequestURI(eventsTargetURL)
+	fmt.Printf("DownstreamReq: %")
+
+	reqURL, err := url.ParseRequestURI(v2.EventsTargetURL)
+
 	if err != nil {
 		return nil, err
 	}
@@ -53,11 +58,7 @@ func SendEvent(req interface{}, traceHeaders *map[string]string,
 	if err != nil {
 		return nil, err
 	}
-	b, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		return nil, err
-	}
-	fmt.Printf("response: %s", string(b))
+
 	defer resp.Body.Close()
 
 	response := api.SendEventResponse{}
@@ -67,7 +68,7 @@ func SendEvent(req interface{}, traceHeaders *map[string]string,
 		if err != nil {
 			return nil, err
 		}
-
+		fmt.Println("what: " + string(body))
 		result := &api.PublishResponse{}
 
 		err = json.Unmarshal(body, result)
