+++ components/event-service/internal/events/bus/v2/process.go
@@ -1,18 +1,46 @@
 package v2
 
 import (
+	"fmt"
+
 	"github.com/kyma-project/kyma/components/event-service/internal/events/api"
-	"github.com/kyma-project/kyma/components/event-service/internal/events/bus"
 )
+
+type configurationData struct {
+	SourceID string
+}
+
+//Conf Event-Service configuration data
+var Conf *configurationData
+
+//EventsTargetURL Event-Service contains target URL
+var EventsTargetURL string
+
+// Init should be used to initialize the "source" related configuration data
+func Init(sourceID string, targetURL string) {
+	Conf = &configurationData{
+		SourceID: sourceID,
+	}
+
+	EventsTargetURL = targetURL
+}
+
+//CheckConf assert the configuration initialization
+func CheckConf() (err error) {
+	if Conf == nil {
+		return fmt.Errorf("configuration data not initialized")
+	}
+	return nil
+}
 
 // AddSource adds the "source" related data to the incoming request
 func AddSource(parameters *api.PublishEventParametersV3) (resp *api.SendEventParametersV3, err error) {
-	if err := bus.CheckConf(); err != nil {
+	if err := CheckConf(); err != nil {
 		return nil, err
 	}
 
 	sendRequest := api.SendEventParametersV3{
-		SourceID:            bus.Conf.SourceID, // enrich the event with the sourceID
+		SourceID:            Conf.SourceID, // enrich the event with the sourceID
 		EventType:           parameters.EventRequestV3.EventType,
 		EventTypeVersion:    parameters.EventRequestV3.EventTypeVersion,
 		EventID:             parameters.EventRequestV3.EventID,
